+++ R/aqiCategories.R
@@ -4,7 +4,7 @@
 #'
 #' @param x Vector or matrix of PM2.5 values or an \emph{mts_monitor} object.
 #' @param pollutant EPA AQS criteria pollutant.
-#' @param NAAQS Version of NAAQS levels to use. See Note.
+#' @param NAAQS User provided NAAQS levels to use.
 #' @param conversionArray Array of six text or other values to return instead of integers.
 #'
 #' @return A vector or matrix of AQI category indices in the range 1:6.
@@ -19,6 +19,10 @@
 #' \code{conversionArray} parameter can be used to convert these integers into
 #' whatever is specified in the first six elements of \code{conversionArray}. A
 #' typical usage would be: \code{conversionArray = US_AQI$names_eng}.
+#'
+#' By default, an appropriate set of NAAQS levels will be chosen for each
+#' \code{pollutant}. Users can override these values by providing an alternate
+#' set of breaks, \emph{e.g.}, \code{NAAQS = US_AQI$breaks_PM2.5_24hr_pre_2024}.
 #'
 #' @note
 #' On February 7, 2024, EPA strengthened the National Ambient Air Quality
@@ -53,21 +57,24 @@
 
 aqiCategories <- function(
   x,
-  pollutant = c("PM2.5", "AQI", "CO", "NO", "OZONE", "PM10", "SO2"),
-  NAAQS = c("PM2.5_2024", "PM2.5"),
+  pollutant = c("PM2.5", "CO", "OZONE", "PM10", "AQI"),
+  NAAQS = NULL,
   conversionArray = NULL
 ) {
 
   # ----- Validate parameters --------------------------------------------------
 
   pollutant <- match.arg(pollutant)
-  NAAQS = match.arg(NAAQS)
 
   breaks <- US_AQI[[paste0("breaks_", pollutant)]]
 
-  # Handle the added NAAQS argument
-  if ( pollutant == "PM2.5" && NAAQS == "PM2.5_2024" ) {
-    breaks <- US_AQI$breaks_PM2.5_2024
+  # Use NAAQS if provided and valid
+  if ( !is.null(NAAQS) ) {
+    if ( !is.numeric(NAAQS) || length(NAAQS) != 7 ) {
+      warning("User provided 'NAAQS' must have 7 numeric levels")
+    } else {
+      breaks <- NAAQS
+    }
   }
 
   # ----- Prepare data ---------------------------------------------------------