+++ R/monitor_timeseriesPlot.R
@@ -23,11 +23,15 @@
 #' @param opacity Opacity to use for points. By default, an opacity is chosen based
 #' on the number of points so that trends are highlighted while outliers diminish
 #' in visual importance as the number of points increases.
-#' @param NAAQS Version of NAAQS levels to use. See Note.
+#' @param NAAQS User provided NAAQS levels to use.
 #' @param ... Additional arguments to be passed to \code{graphics::plot.default()}.
 #'
 #' @return No return value. This function is called to draw an air quality
 #' time series plot on the active graphics device.
+#'
+#' By default, an appropriate set of NAAQS levels will be chosen for \code{monitor$meta$pollutant}
+#' Users can override these values by providing an alternate
+#' set of breaks, \emph{e.g.}, \code{NAAQS = US_AQI$breaks_PM2.5_24hr_pre_2024}.
 #'
 #' @note
 #' On February 7, 2024, EPA strengthened the National Ambient Air Quality
@@ -61,14 +65,14 @@
 #'   )
 #' addAQILegend()
 #'
-#' # Standard extras using the updated PM NAAQS
+#' # Standard extras using the pre-2024 NAAQS
 #' Carmel_Valley %>%
 #'   monitor_timeseriesPlot(
 #'     shadedNight = TRUE,
 #'     addAQI = TRUE,
-#'     NAAQS = "PM2.5_2024"
+#'     NAAQS = US_AQI$breaks_PM2.5_24hr_pre_2024
 #'   )
-#' addAQILegend(NAAQS = "PM2.5_2024")
+#' addAQILegend(NAAQS = US_AQI$breaks_PM2.5_24hr_pre_2024)
 #'
 #' # Fancy plot based on pm2.5 values
 #' pm2.5 <- Carmel_Valley$data[,2]
@@ -90,7 +94,7 @@ monitor_timeseriesPlot <- function(
   addAQI = FALSE,
   palette = c("EPA", "subdued", "deuteranopia"),
   opacity = NULL,
-  NAAQS = c("PM2.5_2024", "PM2.5"),
+  NAAQS = NULL,
   ...
 ) {
 
@@ -101,7 +105,6 @@ monitor_timeseriesPlot <- function(
   add <- MazamaCoreUtils::setIfNull(add, FALSE)
   addAQI <- MazamaCoreUtils::setIfNull(addAQI, FALSE)
   palette <- match.arg(palette)
-  NAAQS = match.arg(NAAQS)
 
   # Subset 'monitor' to a single time series
   if ( !is.null(id) ) {